push

+++ 010_randomize.py
@@ -127,15 +127,16 @@ def randomize_two_sample(df):
     rois = ['SUBCORTICAL']
     # Run Randomize
     for roi in rois:
-        print '######################################'
-        print 'Running randomise for roi:', roi
-        qsm_list = [os.path.join(workspace_iron, subject, 'QSM/QSMnorm_MNI1mm_%s.nii.gz' % roi) for subject in population]
-        #print qsm_list
-        stats_dir = mkdir_path(os.path.join(ahba_dir,  'RANDOMISE'))
-        os.chdir(stats_dir)
-        os.system('fslmerge -t concat_%s.nii.gz %s' % (roi, ' '.join(qsm_list)))
-        os.system('randomise -i concat_%s -o randomise_%s -d design_twosample.mat -t design_twosample.con -R'% (roi, roi))
-        os.system('rm -rf *concat*')
+        if not os.path.isfile('randomise_%s_tstat1.nii.gz'%roi):
+            print '######################################'
+            print 'Running randomise for roi:', roi
+            qsm_list = [os.path.join(workspace_iron, subject, 'QSM/QSMnorm_MNI1mm_%s.nii.gz' % roi) for subject in population]
+            #print qsm_list
+            stats_dir = mkdir_path(os.path.join(ahba_dir,  'RANDOMISE'))
+            os.chdir(stats_dir)
+            os.system('fslmerge -t concat_%s.nii.gz %s' % (roi, ' '.join(qsm_list)))
+            os.system('randomise -i concat_%s -o randomise_%s -d design_twosample.mat -t design_twosample.con -R'% (roi, roi))
+            os.system('rm -rf *concat*')
 
 
 def randomize_one_sample(df):
